fix(eth): harden APDU handling against malformed input
Found by the fuzzing harness added in this PR:

- NULL source passed to memcpy() when an EIP-712 struct-impl APDU or a
  zero-length typed-data leaf carries no payload, which apdu_parser() reports
  as a NULL data pointer whenever Lc is 0
- missing bounds and NULL checks across the generic tx parser fields
- zero-length copies while streaming a personal message
- out-of-range integer conversions in the signature parity byte and the ETH2
  public-key status word

+++ src/features/get_eth2_public_key/cmd_get_eth2_public_key.c
@@ -98,7 +98,8 @@ uint16_t handle_get_eth2_public_key(uint8_t p1,
     // Return code will be sent after UI approve/cancel
     error = 0;
 end:
-    return error;
+    // A cx_err_t does not fit a status word; the low half is what the client sees.
+    return (uint16_t) error;
 }
 
 #endif
